third_party/blink/web_tests/html/selectlist/selectlist-typeahead-disabled-option.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<selectlist id="selectlist" tabindex="0">
    <option>Canada</option>
    <option>Spain</option>
    <option disabled>United Arab Emirates</option>
    <option>United States</option>
</selectlist>
<script>
description('Typeahead does not select disabled selectlist option');

window.jsTestIsAsync = true;

function keyDown(key)
{
    if (!window.eventSender)
      debug("No event sender");
    eventSender.keyDown(key);
}

var selectlist = document.getElementById("selectlist");

selectlist.focus();
selectlist.value = "Canada";
// Should skip disabled "United Arab Emirates" <option>
keyDown('U');
shouldBeFalse('selectlist.open');
shouldBeEqualToString('selectlist.value', 'United States');
finishJSTest();
</script>
</body>
